+++ src/features/cron-history/cron-history.entity.ts
@@ -17,7 +17,7 @@ export default class CronHistoryEntity {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = false;
 
-	@PrimaryGeneratedColumn({ type: 'bigint', unsigned: false })
+	@PrimaryGeneratedColumn({ type: 'int', unsigned: false })
 	id!: number;
 
 	@Column('varchar', { nullable: false })
