+++ src/features/log-data/log-data.entity.ts
@@ -33,7 +33,7 @@ export default class LogDataEntity {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@PrimaryGeneratedColumn({ type: 'bigint', unsigned: false })
+	@PrimaryGeneratedColumn({ type: 'int', unsigned: false })
 	id!: number;
 
 	@Column('char', { length: 36, nullable: false })
