+++ src/features/order/order.entity.ts
@@ -36,7 +36,7 @@ export default class OrderEntity extends EntityAbstract {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	@Index('IDX_order_client_id')
 	client_id!: number;
